requirements.txt
Package                   Version
------------------------- ---------
APScheduler               3.6.3
certifi                   2020.6.20
chardet                   3.0.4
click                     7.1.2
colorhash                 1.0.2
configparser              5.0.0
Flask                     1.1.1
Flask-MonitoringDashboard 3.1.0
idna                      2.10
itsdangerous              1.1.0
Jinja2                    2.11.1
MarkupSafe                1.1.1
numpy                     1.19.2
Pillow                    7.2.0
pip                       20.2.2
psutil                    5.7.2
pytz                      2020.1
qrcode                    6.1
requests                  2.24.0
scipy                     1.5.2
setuptools                49.6.0
six                       1.15.0
SQLAlchemy                1.3.19
sqlitedict                1.7.0
tzlocal                   2.1
urllib3                   1.25.10
Werkzeug                  1.0.1
wheel                     0.35.1
